services/trainer/pyproject.toml
[tool.poetry]
name = "trainer"
version = "0.1.0"
description = ""
authors = ["Harry Keightley <harrykeightley@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.12.1"
transformers = "^4.21.2"
datasets = "^2.4.0"
loguru = "^0.6.0"
Flask = "^2.2.2"
gunicorn = "^20.1.0"
pyhumps = "^3.7.3"
google-cloud-storage = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pyright = "^1.1.270"

[tool.pyright]
reportPrivateImportUsage = false

[tool.pytest.ini_options]
markers = [
  "integration: mark a test as integration (deselect with '-m \"not integration\"')",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
